Function telemetry::gather_telemetry

pub fn gather_telemetry(
    port_id: &str, 
    tx: Sender<TelemetryChannelType>, 
    file_buf: Option<BufWriter<File>>
)

Open a serial port, consume it endlessly and send back parsed telemetry messages through a channel

port_id - Name or path to the serial port.

tx - Sender of a channel.

file_buf - Optional file buffer; if specified, messages will also be serialized and written in this file.

This is meant to be run in a dedicated thread.