Struct telemetry::keyed::KeyedCount

pub struct KeyedCount<K> {
    // some fields omitted
}

Count histograms.

A Count histogram simply accumulates the numbers passed with record().

With SerializationFormat::SimpleJson, these histograms are serialized as an object, with keys sorted, in which each field is a number.

Methods

impl<K> KeyedCount<K>

fn new(feature: &Feature, name: String) -> KeyedCount<K>