Struct telemetry::keyed::KeyedLinear

pub struct KeyedLinear<K, T> where T: Flatten {
    // some fields omitted
}

Linear histograms.

Linear histograms classify numeric integer values into same-sized buckets. This type is typically used for percentages, or to store a relatively precise of the amount of resources (time, memory) used by a section.

With SerializationFormat::SimpleJson, these histograms are serialized as an object js { key_1: array_1, key_2: array_2, ... }

where each array_i is an array of numbers, one per bucket, in the numeric order of buckets

Methods

impl<K, T> KeyedLinear<K, T> where K: ToString, T: Flatten

fn new(feature: &Service, name: String, min: u32, max: u32, buckets: usize) -> KeyedLinear<K, T>