Module telemetry::plain

Definition of histograms.

Histograms represent measures on a set of entities known at compile-time. In some cases, the set of entities is not known at compile-time (e.g. plug-ins, dates), in which case you should rather use Keyed histograms.

Structs

Count

Enum

Flag

Ignoring

Linear

Traits

Histogram