Struct telemetry::plain::Count

pub struct Count {
    // some fields omitted
}

Count histograms.

A Count histogram simply accumulates the numbers passed with record(). Count histograms are useful, for instance, to know how many times a feature has been used, or how many times an error has been triggered.

With SerializationFormat::SimpleJson, these histograms are serialized as a plain number.

Methods

impl Count

fn new(feature: &Service, name: String) -> Count