Struct telemetry::plain::Enum

pub struct Enum<K> where K: Flatten {
    // some fields omitted
}

Enumerated histograms.

Enumerated histogram generalize Count histograms to families of keys known at compile-time. They are useful, for instance, to know how often users have picked a specific choice from several, or how many times each kind of error has been triggered, etc.

With SerializationFormat::SimpleJson, these histograms are serialized as an array of numbers, in the order of enum values.

Methods

impl<K> Enum<K> where K: Flatten

fn new(service: &Service, name: String) -> Enum<K>

Create a new Enum histogram with a given name.

Argument name is used as key when processing and exporting the data. Each name must be unique to the Service.

Panics

If name is already used by another histogram in service.