Struct telemetry::plain::Flag

pub struct Flag {
    // some fields omitted
}

Flag histograms.

This histogram has only two states. Until the first call to record(), it is unset. Once record() has been called once, it is set and won't change anymore. This type is useful if you need to track whether a feature was ever used during a session.

With SerializationFormat::SimpleJson, these histograms are serialized as a plain number 0 (unset)/1 (set).

Methods

impl Flag

fn new(service: &Service, name: String) -> Flag

Create a new Flag histogram with a given name.

Argument name is used as key when processing and exporting the data. Each name must be unique to the Service.

Panics

If name is already used by another histogram in service.