Trait telemetry::plain::Histogram

pub trait Histogram<T> {
    fn record_cb<F>(&self, _: F) where F: FnOnce() -> Option<T>;

    fn record(&self, value: T) { ... }
}

A plain histogram.

Required Methods

fn record_cb<F>(&self, _: F) where F: FnOnce() -> Option<T>

Record a value in this histogram, as provided by a callback.

If the service is currently inactive, this is a noop.

If the callback returns None, no value is recorded.

Provided Methods

fn record(&self, value: T)

Record a value in this histogram.

If the service is currently inactive, this is a noop.