Struct telemetry::Service

pub struct Service {
    // some fields omitted
}

Methods

impl Service

The Telemetry service.

Panics

The service will panic if an attempt is made to register two histograms with the same key.

fn new() -> Service

fn to_json(&self, format: SerializationFormat, sender: Sender<(Json, Json)>)

Serialize all histograms as json, in a given format.

Returns a pair with plain histograms/keyed histograms.

fn set_active(&self, value: bool)

Make the service (in)active.

Any data recorded on a histogram while the service is inactive will be ignored.

fn is_active(&self) -> bool