Crate telemetry

Re-exports

pub use serial;

Modules

alarm
Utilities related to alarms

control
Structures to represent control messages

structures
Structures to represent telemetry messages

Functions

display_message
Helper to display telemetry messages

gather_telemetry
Open a serial port, consume it endlessly and send back parsed telemetry messages through a channel

gather_telemetry_from_file
Open a file containing serialized telemetry data, read it and send back parsed telemetry messages through a channel

Type Definitions

TelemetryChannelType